fix(sessions): address coverage and package shape review
